Split TypeDefs into new files. #10
